Added authentication and authorization , updated goal REST API

+++ backend/controllers/goalController.js
@@ -1,25 +1,27 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 
 // Get goals GET /api/goals Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals)
 })
 
 
 // Set goals POST /api/goals Private
-const setGoals = asyncHandler(async (req, res) => {
-    if (!req.body) {
+const setGoal = asyncHandler(async (req, res) => {
+    if (!req.body.text) {
         res.status(404)
         throw new Error('Please add a text field')
     }
 
     const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id,
     })
 
     res.status(200).json(goal);
@@ -28,12 +30,24 @@ const setGoals = asyncHandler(async (req, res) => {
 
 // Update goals PUT /api/goals/:id Private
 const updateGoal = asyncHandler(async (req, res) => {
+    const user = await User.findById(req.user.id)
+    // check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found!!')
+    }
+
     const goal = await Goal.findById(req.params.id)
 
     if (!goal) {
         res.status(400)
         throw new Error('Goal not Found!!')
     }
+    if (goal.user.toString() != user.id) {
+        res.status(401)
+        throw new Error('User not Autorized')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
     })
@@ -44,11 +58,22 @@ const updateGoal = asyncHandler(async (req, res) => {
 
 // Delete goals DELETE /api/goals/:id Private
 const deleteGoal = asyncHandler(async (req, res) => {
-    const goal = await Goal.findById(req.params.id)
+    const user = await User.findById(req.user.id)
+    // check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found!!')
+    }
 
+    const goal = await Goal.findById(req.params.id)
+    
     if (!goal) {
         res.status(400)
-        throw new Error('Goal Not Found!!')
+        throw new Error('Goal not Found!!')
+    }
+    if (goal.user.toString() != user.id) {
+        res.status(401)
+        throw new Error('User not Autorized')
     }
 
     await goal.deleteOne()
@@ -58,7 +83,7 @@ const deleteGoal = asyncHandler(async (req, res) => {
 
 module.exports = {
     getGoals,
-    setGoals,
+    setGoal,
     updateGoal,
     deleteGoal
 }
